move multislider source
