+++ src/viser/client/src/components/MultiSliderPrimitive/MultiSlider.styles.tsx
